additional fix moments for tests, helpers and data
