+++ src/test/java/data/TestDataGeneration.java
@@ -10,15 +10,19 @@ public class TestDataGeneration {
 	public static final int PAGE = 1;
 	public static final int PER_PAGE = 10;
 
-	public static EntityRequest generateEntity() {
+	public static AdditionRequest generateAddition() {
 		AdditionRequest addition = new AdditionRequest();
 		addition.setAdditionalInfo(faker.lorem().sentence());
 		addition.setAdditionalNumber(faker.number().numberBetween(1, 1000));
+		return addition;
+	}
+
+	public static EntityRequest generateEntity() {
 		EntityRequest entity = new EntityRequest();
 		entity.setTitle(faker.company().name());
 		entity.setVerified(faker.bool().bool());
 		entity.setImportantNumbers(Arrays.asList(faker.number().randomDigit(), faker.number().randomDigit()));
-		entity.setAddition(addition);
+		entity.setAddition(generateAddition());
 		return entity;
 	}
