+++ src/test/java/helpers/ApiEndpoints.java
@@ -1,6 +1,6 @@
 package helpers;
 
-public class APIEndpoints {
+public class ApiEndpoints {
 	public static final String BASE_URI = "http://localhost:8080/";
 	public static final String CREATE = "/api/create";
 	public static final String GET = "/api/get/{id}";
